HW1/src/main/java/MyArrayList.java
import java.util.Arrays;

public class MyArrayList<T> implements MyList<T> {
    private Object[] array;
    private int size;

    public MyArrayList() {
        this.array = new Object[10];
        this.size = 0;
    }

    public MyArrayList(int initCapacity) {
        this.array = new Object[initCapacity];
        this.size = 0;
    }

    public MyArrayList(Object[] elements) {
        this.array = new Object[elements.length * 2];
        for (int i = 0; i < elements.length; i++) {
            this.array[i] = elements[i];
        }
        this.size = elements.length;
    }

    public void add(T element) {
        if (size == array.length) {
            Object[] newArray = new Object[array.length * 2];
            for (int i = 0; i < size; i++) {
                newArray[i] = array[i];
            }
            array = newArray;
        }
        array[size] = element;
        size++;
    }

    public void remove(int index) {
        if (index < 0 || index >= size) {
            throw new IndexOutOfBoundsException();
        }

        for (int i = index; i < size - 1; i++) {
            array[i] = array[i + 1];
        }
        array[size - 1] = null;
        size--;
    }

    public T get(int index) {
        if (index < 0 || index >= size) {
            throw new IndexOutOfBoundsException();
        }
        return (T) array[index];
    }

    public int size() {
        return size;
    }

    @Override
    public String toString() {
        StringBuilder sb = new StringBuilder();
        sb.append("[");
        for (int i = 0; i < size(); i++) {
            sb.append(get(i));
            if (i < size() - 1) {
                sb.append(", ");
            }
        }
        sb.append("]");
        return sb.toString();
    }
}